+++ app.py
@@ -1,20 +1,28 @@
 import streamlit as st
+from utils import util
 import pandas as pd
-import matplotlib.pyplot as plt
-import seaborn as sns
-from utils import login
+import numpy as np
 
-st.set_page_config(
-    page_title="Soccer Central",
-    page_icon="📊",
-    layout="wide",
-    initial_sidebar_state="expanded"
-)
+util.setup_page("Soccer Central")
 
-login.generarLogin()
+st.header(":orange[Home]", divider=True)
 
-if "usuario" not in st.session_state:
-    st.stop()
+a, b, c ,d  = st.columns(4)
 
-st.header(" :orange[Inicio]", divider=True)
+a.metric("Temperature", "30°F", "-9°F", border=True)
+b.metric("Wind", "4 mph", "2 mph", border=True)
 
+c.metric("Humidity", "77%", "5%", border=True)
+d.metric("Pressure", "30.34 inHg", "-2 inHg", border=True)
+
+st.divider()
+e, f  = st.columns(2)
+chart_data = pd.DataFrame(np.random.randn(20, 3), columns=["a", "b", "c"])
+
+e.area_chart(chart_data)
+
+#st.divider()
+
+chart_data = pd.DataFrame(np.random.randn(20, 3), columns=["a", "b", "c"])
+
+f.bar_chart(chart_data)